ai_assistant V1.1.0/requirements.txt
PyQt5>=5.15.0
tqdm>=4.65.0
diffusers>=0.24.0
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0